requirements.txt
matplotlib
numpy
pytorch
sklearn